Feature/candidate (#684)
* update

* update

* update

* update

* 更新添加候选人信息的后台代码

* candidate部分css和js提交

* 页面提交

* 网页部分样式修改

* 添加填写候选人信息的模态框

* 页面样式修改

* 填写内容提交

* 页面数据展示

* 数据展示

* update

* 页面数据

* 解决冲突

* 为投票页面添加多语言

* 添加公钥验证

* update

* 区块高度和时间计算

* 倒计时计算

* fixed bug

* 页面样式修改

* support https

* 页面样式，内容调整

* 页面样式调整

* 部分页面内容提交

* 页面样式细化

* 手机端页面适应，样式调整

* 页面信息调整

* 页面内容修改

* 修改数据结构，删除非必要字段

* 页面修改

* update

* update

* update

* rename

* update

* 页面链接完善

* 样式微调

+++ NeoWeb/Models/ICO2Models.cs
@@ -7,13 +7,14 @@ using System.Threading.Tasks;
 
 namespace NeoWeb.Models
 {
+    //Abandoned
     public class ICO2
     {
         [Key]
         public string NeoAddress { get; set; }
 
         [Required]
-        public decimal GiveBackCNY { get; set; }
+        public double GiveBackCNY { get; set; }
 
         [NotMapped]
         public Choose Choose { get; set; }
